rearranged packages to include service and secondservice into gateway customfilters package, as they belong to that article
