--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/secondservice/SecondServiceApplication.java
@@ -1,15 +1,0 @@
-package com.baeldung.secondservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.PropertySource;
-
-@SpringBootApplication
-@PropertySource("classpath:secondservice-application.properties")
-public class SecondServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SecondServiceApplication.class, args);
-    }
-
-}
