--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/secondservice/web/SecondServiceRestController.java
@@ -1,18 +1,0 @@
-package com.baeldung.secondservice.web;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import reactor.core.publisher.Mono;
-
-@RestController
-public class SecondServiceRestController {
-
-    @GetMapping("/resource/language")
-    public Mono<ResponseEntity<String>> getResource() {
-        return Mono.just(ResponseEntity.ok()
-            .body("es"));
-
-    }
-}
