--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/service/ServiceApplication.java
@@ -1,15 +1,0 @@
-package com.baeldung.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.PropertySource;
-
-@SpringBootApplication
-@PropertySource("classpath:service-application.properties")
-public class ServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ServiceApplication.class, args);
-    }
-
-}
