--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/springcloudgateway/customfilters/config/WebClientConfig.java
@@ -1,16 +1,0 @@
-package com.baeldung.springcloudgateway.customfilters.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    WebClient client() {
-        return WebClient.builder()
-            .build();
-    }
-
-}
