--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/springcloudgateway/customfilters/filters/factories/LoggingGatewayFilterFactory.java
@@ -1,85 +1,0 @@
-package com.baeldung.springcloudgateway.customfilters.filters.factories;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.OrderedGatewayFilter;
-import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.stereotype.Component;
-
-import reactor.core.publisher.Mono;
-
-@Component
-public class LoggingGatewayFilterFactory extends AbstractGatewayFilterFactory<LoggingGatewayFilterFactory.Config> {
-
-    final Logger logger = LoggerFactory.getLogger(LoggingGatewayFilterFactory.class);
-
-    public static final String BASE_MSG = "baseMessage";
-    public static final String PRE_LOGGER = "preLogger";
-    public static final String POST_LOGGER = "postLogger";
-
-    public LoggingGatewayFilterFactory() {
-        super(Config.class);
-    }
-
-    @Override
-    public List<String> shortcutFieldOrder() {
-        return Arrays.asList(BASE_MSG, PRE_LOGGER, POST_LOGGER);
-    }
-
-    @Override
-    public GatewayFilter apply(Config config) {
-        return new OrderedGatewayFilter((exchange, chain) -> {
-            if (config.isPreLogger())
-                logger.info("Pre GatewayFilter logging: " + config.getBaseMessage());
-            return chain.filter(exchange)
-                .then(Mono.fromRunnable(() -> {
-                    if (config.isPostLogger())
-                        logger.info("Post GatewayFilter logging: " + config.getBaseMessage());
-                }));
-        }, 1);
-    }
-
-    public static class Config {
-        private String baseMessage;
-        private boolean preLogger;
-        private boolean postLogger;
-
-        public Config() {
-        };
-
-        public Config(String baseMessage, boolean preLogger, boolean postLogger) {
-            super();
-            this.baseMessage = baseMessage;
-            this.preLogger = preLogger;
-            this.postLogger = postLogger;
-        }
-
-        public String getBaseMessage() {
-            return this.baseMessage;
-        }
-
-        public boolean isPreLogger() {
-            return preLogger;
-        }
-
-        public boolean isPostLogger() {
-            return postLogger;
-        }
-
-        public void setBaseMessage(String baseMessage) {
-            this.baseMessage = baseMessage;
-        }
-
-        public void setPreLogger(boolean preLogger) {
-            this.preLogger = preLogger;
-        }
-
-        public void setPostLogger(boolean postLogger) {
-            this.postLogger = postLogger;
-        }
-    }
-}
