--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/springcloudgateway/customfilters/filters/global/FirstPreLastPostGlobalFilter.java
@@ -1,31 +1,0 @@
-package com.baeldung.springcloudgateway.customfilters.filters.global;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-import org.springframework.cloud.gateway.filter.GlobalFilter;
-import org.springframework.core.Ordered;
-import org.springframework.stereotype.Component;
-import org.springframework.web.server.ServerWebExchange;
-
-import reactor.core.publisher.Mono;
-
-@Component
-public class FirstPreLastPostGlobalFilter implements GlobalFilter, Ordered {
-
-    final Logger logger = LoggerFactory.getLogger(FirstPreLastPostGlobalFilter.class);
-
-    @Override
-    public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-        logger.info("First Pre Global Filter");
-        return chain.filter(exchange)
-            .then(Mono.fromRunnable(() -> {
-                logger.info("Last Post Global Filter");
-            }));
-    }
-
-    @Override
-    public int getOrder() {
-        return -1;
-    }
-}
