--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/springcloudgateway/customfilters/filters/global/LoggingGlobalFiltersConfigurations.java
@@ -1,25 +1,0 @@
-package com.baeldung.springcloudgateway.customfilters.filters.global;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.cloud.gateway.filter.GlobalFilter;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import reactor.core.publisher.Mono;
-
-@Configuration
-public class LoggingGlobalFiltersConfigurations {
-
-    final Logger logger = LoggerFactory.getLogger(LoggingGlobalFiltersConfigurations.class);
-
-    @Bean
-    public GlobalFilter postGlobalFilter() {
-        return (exchange, chain) -> {
-            return chain.filter(exchange)
-                .then(Mono.fromRunnable(() -> {
-                    logger.info("Global Post Filter executed");
-                }));
-        };
-    }
-}
