--- spring-cloud/spring-cloud-gateway/src/main/java/com/baeldung/springcloudgateway/customfilters/filters/global/LoggingGlobalPreFilter.java
@@ -1,22 +1,0 @@
-package com.baeldung.springcloudgateway.customfilters.filters.global;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-import org.springframework.cloud.gateway.filter.GlobalFilter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.server.ServerWebExchange;
-
-import reactor.core.publisher.Mono;
-
-@Component
-public class LoggingGlobalPreFilter implements GlobalFilter {
-
-    final Logger logger = LoggerFactory.getLogger(LoggingGlobalPreFilter.class);
-
-    @Override
-    public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-        logger.info("Global Pre Filter executed");
-        return chain.filter(exchange);
-    }
-}
