--- spring-cloud/spring-cloud-gateway/src/test/java/com/baeldung/secondservice/SecondServiceIntegrationTest.java
@@ -1,26 +1,0 @@
-package com.baeldung.secondservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
-import org.springframework.test.web.reactive.server.WebTestClient;
-
-import com.baeldung.secondservice.web.SecondServiceRestController;
-
-@WebFluxTest(SecondServiceRestController.class)
-public class SecondServiceIntegrationTest {
-
-    @Autowired
-    private WebTestClient webClient;
-
-    @Test
-    public void whenResourceLanguageEndpointCalled_thenRetrievesSpanishLanguageString() throws Exception {
-        this.webClient.get()
-            .uri("/resource/language")
-            .exchange()
-            .expectStatus()
-            .isOk()
-            .expectBody(String.class)
-            .isEqualTo("es");
-    }
-}
