--- spring-cloud/spring-cloud-gateway/src/test/java/com/baeldung/service/SpringContextTest.java
@@ -1,12 +1,0 @@
-package com.baeldung.service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = ServiceApplication.class)
-public class SpringContextTest {
-
-    @Test
-    public void whenSpringContextIsBootstrapped_thenNoExceptions() {
-    }
-}
