+++ spring-cloud/spring-cloud-gateway/src/test/resources/logback-test.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
     <appender name="LISTAPPENDER"
-        class="com.baeldung.springcloudgateway.customfilters.utils.LoggerListAppender">
+        class="com.baeldung.springcloudgateway.customfilters.gatewayapp.utils.LoggerListAppender">
     </appender>
     <root level="info">
         <appender-ref ref="LISTAPPENDER" />
